Inspect AST rather than deparse/split RHS
The `?` operator previously deparsed the rhs and split on all colons,
this caused bugs in strings with colons, function calls under namespaces
and meant that multiple statements could not be chained.

Instead, the abstract syntax tree (AST) of the RHS is inspected. As `:`
is a function, its lhs and rhs are branches of the syntax tree
once parsed by the R interpreter. By inspecting the AST and assigning
the lhs and rhs of the `:` call to a vector the expression in the RHS
can be split without interference from colons not acting as operators